Feat: Plate thrower now uses the same base throwing script as the bookshelf
