+++ Assets/Scenes/Mansion.unity
@@ -27087,6 +27087,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 1059055258}
+  - {fileID: 1953511703}
   m_Father: {fileID: 1413777081}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &1641039336
@@ -29113,6 +29114,86 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 0.2, y: 3, z: 1}
   m_Center: {x: 0, y: 1.5, z: 0}
+--- !u!1001 &1706888569
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 1640203345}
+    m_Modifications:
+    - target: {fileID: 335559059758086393, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -6.225
+      objectReference: {fileID: 0}
+    - target: {fileID: 335559059758086393, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 1.298
+      objectReference: {fileID: 0}
+    - target: {fileID: 335559059758086393, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -4.469
+      objectReference: {fileID: 0}
+    - target: {fileID: 335559059758086393, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 335559059758086393, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 335559059758086393, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0.7071068
+      objectReference: {fileID: 0}
+    - target: {fileID: 335559059758086393, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0.7071068
+      objectReference: {fileID: 0}
+    - target: {fileID: 335559059758086393, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: -90
+      objectReference: {fileID: 0}
+    - target: {fileID: 335559059758086393, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 180
+      objectReference: {fileID: 0}
+    - target: {fileID: 335559059758086393, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 538986359465493339, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: InteractEvent.m_PersistentCalls.m_Calls.Array.data[0].m_Target
+      value: 
+      objectReference: {fileID: 1953511713}
+    - target: {fileID: 538986359465493339, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: InteractEvent.m_PersistentCalls.m_Calls.Array.data[0].m_TargetAssemblyTypeName
+      value: PlateThrower, Assembly-CSharp
+      objectReference: {fileID: 0}
+    - target: {fileID: 3641300511731537466, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+        type: 3}
+      propertyPath: m_Name
+      value: Kitchen Horizontal Cabinet Variant
+      objectReference: {fileID: 0}
+    m_RemovedComponents:
+    - {fileID: 6014817620437526480, guid: b0ed7f06880528d4eb097b7cf282e7fa, type: 3}
+    m_RemovedGameObjects:
+    - {fileID: 9094850496847082411, guid: b0ed7f06880528d4eb097b7cf282e7fa, type: 3}
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: b0ed7f06880528d4eb097b7cf282e7fa, type: 3}
 --- !u!1 &1707353351
 GameObject:
   m_ObjectHideFlags: 0
@@ -32088,6 +32169,24 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: b71b0f8d2f881c743a85228f3030d1cb, type: 3}
+--- !u!4 &1953511703 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 335559059758086393, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+    type: 3}
+  m_PrefabInstance: {fileID: 1706888569}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &1953511713 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 5432911223318634606, guid: b0ed7f06880528d4eb097b7cf282e7fa,
+    type: 3}
+  m_PrefabInstance: {fileID: 1706888569}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: de5b99caf511a1349b100483a2ac4460, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &1957385119
 PrefabInstance:
   m_ObjectHideFlags: 0
