+++ Assets/Scripts/Interactable/Interactions/Thrower/CabinetPlateThrower.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 38c6ab7c35acaed4fa2b88f02a4d2efd
+guid: de5b99caf511a1349b100483a2ac4460
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
